test: 역 삭제 통합테스트
