+++ src/main/java/subway/service/line/LineDao.java
@@ -1,0 +1,5 @@
+package subway.service.line;
+
+public interface LineDao {
+
+}
